# liminalis/docs/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeStockRequest": {
        "properties": {
          "lang": {
            "default": "zh",
            "title": "Lang",
            "type": "string"
          },
          "mode": {
            "default": "fast",
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "default": "价值投资分析",
            "title": "Query",
            "type": "string"
          },
          "stock": {
            "default": "",
            "title": "Stock",
            "type": "string"
          },
          "stock_code": {
            "default": "",
            "title": "Stock Code",
            "type": "string"
          },
          "use_cache": {
            "default": true,
            "title": "Use Cache",
            "type": "boolean"
          }
        },
        "title": "AnalyzeStockRequest",
        "type": "object"
      },
...
      "CreateJournalEntryRequest": {
        "properties": {
          "action": {
            "default": "",
            "title": "Action",
            "type": "string"
          },
          "emotion": {
            "default": "",
            "title": "Emotion",
            "type": "string"
          },
          "entry_type": {
            "title": "Entry Type",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "review_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review At"
          },
          "thesis_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Id"
          }
        },
        "required": [
          "entry_type"
        ],
        "title": "CreateJournalEntryRequest",
        "type": "object"
      },
...
      "CreateReviewRequest": {
        "properties": {
          "decision": {
            "default": "no_action",
            "title": "Decision",
            "type": "string"
          },
          "discipline_notes": {
            "default": "",
            "title": "Discipline Notes",
            "type": "string"
          },
          "evidence_update": {
            "default": "",
            "title": "Evidence Update",
            "type": "string"
          },
          "next_action": {
            "default": "",
            "title": "Next Action",
            "type": "string"
          },
          "thesis_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Id"
          },
          "thesis_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Valid"
          },
          "valuation_update": {
            "default": "",
            "title": "Valuation Update",
            "type": "string"
          }
        },
        "title": "CreateReviewRequest",
        "type": "object"
      },
...
      "CreateWatchItemRequest": {
        "properties": {
          "condition": {
            "default": "",
            "title": "Condition",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "kind": {
            "default": "manual",
            "title": "Kind",
            "type": "string"
          },
          "priority": {
            "default": 0,
            "title": "Priority",
            "type": "integer"
          },
          "thesis_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateWatchItemRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateWatchItemStatusRequest": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateWatchItemStatusRequest",
        "type": "object"
      },
      "UpsertThesisRequest": {
        "properties": {
          "confidence": {
            "default": "medium",
            "title": "Confidence",
            "type": "string"
          },
          "core_thesis": {
            "default": "",
            "title": "Core Thesis",
            "type": "string"
          },
          "counter_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Counter Evidence",
            "type": "array"
          },
          "disconfirming_signals": {
            "items": {
              "type": "string"
            },
            "title": "Disconfirming Signals",
            "type": "array"
          },
          "expected_holding_period": {
            "default": "",
            "title": "Expected Holding Period",
            "type": "string"
          },
          "margin_of_safety": {
            "default": "",
            "title": "Margin Of Safety",
            "type": "string"
          },
          "source_report_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Report Id"
          },
          "status": {
            "default": "watchlist",
            "title": "Status",
            "type": "string"
          },
          "stock_code": {
            "default": "",
            "title": "Stock Code",
            "type": "string"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name"
          },
          "supporting_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          }
        },
        "title": "UpsertThesisRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Liminalis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/invest/analyze-stock": {
      "post": {
        "operationId": "post_analyze_stock_api_invest_analyze_stock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Analyze Stock",
        "tags": [
          "invest"
        ]
      }
    },
    "/api/invest/journal/{stock_code}": {
      "post": {
        "operationId": "post_journal_entry_api_invest_journal__stock_code__post",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJournalEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Journal Entry Api Invest Journal  Stock Code  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Journal Entry",
        "tags": [
          "invest"
        ]
      }
    },
    "/api/invest/report/{stock_code}": {
      "get": {
        "operationId": "get_report_api_invest_report__stock_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Report Api Invest Report  Stock Code  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "invest"
        ]
      }
    },
...
    "/api/invest/reviews/{stock_code}": {
      "post": {
        "operationId": "post_review_api_invest_reviews__stock_code__post",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Review Api Invest Reviews  Stock Code  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Review",
        "tags": [
          "invest"
        ]
      }
    },
...
    "/api/invest/thesis/{stock_code}": {
      "put": {
        "operationId": "put_invest_thesis_api_invest_thesis__stock_code__put",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertThesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Invest Thesis Api Invest Thesis  Stock Code  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Invest Thesis",
        "tags": [
          "invest"
        ]
      }
    },
    "/api/invest/watch-items/{item_id}": {
      "delete": {
        "operationId": "delete_watch_item_api_invest_watch_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Watch Item Api Invest Watch Items  Item Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watch Item",
        "tags": [
          "invest"
        ]
      },
      "patch": {
        "operationId": "patch_watch_item_status_api_invest_watch_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWatchItemStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Watch Item Status Api Invest Watch Items  Item Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Watch Item Status",
        "tags": [
          "invest"
        ]
      }
    },
    "/api/invest/watch-items/{stock_code}": {
      "post": {
        "operationId": "post_watch_item_api_invest_watch_items__stock_code__post",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWatchItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Watch Item Api Invest Watch Items  Stock Code  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Watch Item",
        "tags": [
          "invest"
        ]
      }
    },
    "/api/invest/workspace/{stock_code}": {
      "get": {
        "operationId": "get_workspace_api_invest_workspace__stock_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Workspace Api Invest Workspace  Stock Code  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "invest"
        ]
      }
    },
...
  }
}